Add a dependency on `latest_additions`
`pub_lib` embeds `latest_tool` on the front page (`index.html`) both in
'Latest Additions' and 'Recommended Research'. While this would
previously work using the core `latest_tool` (although without all the
new visuals), `latest_additions` is included by default so should be
used and the plan is to remove the core `latest_tool` entirely
(eprints/eprints3.5#230).

As such this ensures that the dependency on `latest_additions` is
codified in this flavour, while you can do without it in a flavour like
`zero`, its results are a big part of the new `pub_lib` frontpage so it
should be included by default.